Added passwords and names to lobbies
Clients can now name lobbies and add passwords when creating. When joining lobbies they are listed by name. When an attempt is made to join a lobby the correct password must be entered to join.

I also made some cosmetic changes to the files I was editing, just for the sake of consistent style. Sorry if this annoys anyone.

Also made slight change to requestLobbies and createLobbies methods in server code. This was to allow for storage of passwords and lobby names. The key in the lobbies data structure is now the lobby name (lobby_id).